requirements.txt
asgiref==3.3.1
astroid==2.3.1
autopep8==1.4.4
colorama==0.4.4
Django==3.1.7
django-multiselectfield==0.1.12
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
Pillow==8.1.2
pycodestyle==2.5.0
pylint==2.4.2
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
virtualenv==16.7.6
wrapt==1.11.2
